Update locate.c
To compile under Ubuntu 24.

+++ codebase/analysis/src.bin/cdf/istp_text.1.24/locate.c
@@ -180,7 +180,7 @@ struct file_list *locate_files(char *fpath,char *prefix,
   
   for (yr=syr;yr<=eyr;yr++) {
     
-    sprintf(dir,"%s/%.4d",fpath,(double) yr);
+    sprintf(dir,"%s/%.4d",fpath, yr);
    
     status=scandir(dir,&dlist,dsel,alphasort);
